move to numpy directory
