+++ crazy_rl/__init__.py
@@ -1,2 +1,2 @@
 """Crazy RL: A Reinforcement Learning Framework for Crazyflie Quadcopters."""
-import crazy_rl.multi_agent
+import crazy_rl.multi_agent.numpy
