--- crazy_rl/multi_agent/__init__.py
@@ -1,6 +1,0 @@
-"""Multi-agent environments for Crazy RL."""
-import crazy_rl.multi_agent.catch
-import crazy_rl.multi_agent.circle
-import crazy_rl.multi_agent.escort
-import crazy_rl.multi_agent.hover
-import crazy_rl.multi_agent.surround
