+++ crazy_rl/multi_agent/numpy/__init__.py
@@ -1,0 +1,6 @@
+"""Multi-agent environments for Crazy RL."""
+import crazy_rl.multi_agent.numpy.catch
+import crazy_rl.multi_agent.numpy.circle
+import crazy_rl.multi_agent.numpy.escort
+import crazy_rl.multi_agent.numpy.hover
+import crazy_rl.multi_agent.numpy.surround
