+++ crazy_rl/multi_agent/numpy/catch/__init__.py
@@ -1,2 +1,1 @@
 """Catch environment for multi-agent reinforcement learning."""
-from crazy_rl.multi_agent.catch.catch import Catch
