+++ crazy_rl/multi_agent/numpy/circle/__init__.py
@@ -1,2 +1,1 @@
 """Circle environment for multi-agent reinforcement learning."""
-from crazy_rl.multi_agent.circle.circle import Circle
