+++ crazy_rl/multi_agent/numpy/circle/circle.py
@@ -7,7 +7,7 @@ import numpy as np
 from gymnasium import spaces
 from pettingzoo.test.parallel_test import parallel_api_test
 
-from crazy_rl.multi_agent.base_parallel_env import BaseParallelEnv
+from crazy_rl.multi_agent.numpy.base_parallel_env import BaseParallelEnv
 
 
 class Circle(BaseParallelEnv):
