+++ crazy_rl/multi_agent/numpy/escort/__init__.py
@@ -1,2 +1,1 @@
 """Escort environment for multi-agent reinforcement learning."""
-from crazy_rl.multi_agent.escort.escort import Escort
