+++ crazy_rl/multi_agent/numpy/hover/__init__.py
@@ -1,2 +1,1 @@
 """Hover environment for multi-agent reinforcement learning."""
-from crazy_rl.multi_agent.hover.hover import Hover
