+++ crazy_rl/multi_agent/numpy/surround/__init__.py
@@ -1,2 +1,2 @@
 """Surround environment for multi-agent reinforcement learning."""
-from crazy_rl.multi_agent.surround.surround import Surround
+from crazy_rl.multi_agent.numpy.surround.surround import Surround
